Divided the init command out into a new component. Build has also been moved into this directory. The reasoning behind this is that the init command seems to be the only command that uses these functions, and this might also make a 'build' command easy to implement going forward.
