+++ src/lib.rs
@@ -11,7 +11,6 @@ extern crate serde_json;
 extern crate toml;
 
 pub mod bindgen;
-pub mod build;
 pub mod command;
 pub mod context;
 pub mod emoji;
